Move the test case to the low-level client tests
